Wait for Setup Assistant to be Completed (#30)
Another friendly suggestion. I ran into an issue when using this script deployed to "all devices" on a fresh system. If it runs before setup assistant is completed and the first user logs it will steal the secure token for the admin account from the main account and cause all kinds of issues (FileVault namely won't be successful for main user account login).