Update installApp3.01-GoogleChrome.sh
Updates the web URL to support the use of the universal binary for Google Chrome Enterprise on macOS in .DMG format. The universal binary supports both x86 and ARM (Apple M chips). The use of the enterprise edition is relevant here since these Intune scripts will be primarily used in enterprise environments.